Makes more sense to use int over size_t

+++ src/GUI/PC_Gui.hpp
@@ -127,7 +127,7 @@ private:
 	
 public:
 	// Maps user id -> name
-	std::map<size_t, gchar*> users;
+	std::map<int, gchar*> users;
 	
 // Constructor, destructor, and initializer
 public:
